Use cffi 1.X features to build native module for faster import
